Initial basemap model and selector implementation
